Add delta renderer + fix all clippy warnings
Delta renderer only redraws terminal cells that changed between frames,
dramatically reducing escape sequence output for animations where most
cells are static (fire upper half, starfield background, etc).

Key for large terminals (200x44+) through tmux where PTY processing
of redundant escape sequences was causing jitter/low FPS.

- Falls back to full frame when >60% of cells changed
- Invalidates on resize/mode change
- Also fixed all clippy warnings (too_many_arguments, matches!, etc.)

+++ src/animations/aurora.rs
@@ -1,5 +1,5 @@
-use crate::render::Canvas;
 use super::Animation;
+use crate::render::Canvas;
 use noise::{NoiseFn, Perlin};
 
 /// Aurora borealis effect
@@ -16,7 +16,9 @@ impl Aurora {
 }
 
 impl Animation for Aurora {
-    fn name(&self) -> &str { "aurora" }
+    fn name(&self) -> &str {
+        "aurora"
+    }
 
     fn update(&mut self, canvas: &mut Canvas, _dt: f64, time: f64) {
         let w = canvas.width as f64;
@@ -30,7 +32,7 @@ impl Animation for Aurora {
         for i in 0..80 {
             let sx = ((star_seed.wrapping_mul(i * 7 + 3)) % canvas.width as u64) as usize;
             let sy = ((star_seed.wrapping_mul(i * 13 + 7)) % (canvas.height as u64 / 3)) as usize;
-            let twinkle = ((t * 2.0 + i as f64 * 0.7).sin() * 0.5 + 0.5) as f64;
+            let twinkle = (t * 2.0 + i as f64 * 0.7).sin() * 0.5 + 0.5;
             if twinkle > 0.3 {
                 let b = (twinkle * 120.0) as u8;
                 canvas.set_colored(sx, sy, twinkle * 0.4, b, b, (b as u16 + 30).min(255) as u8);
@@ -46,9 +48,12 @@ impl Animation for Aurora {
                 let speed = 0.8 + layer as f64 * 0.2;
 
                 let n1 = self.noise.get([fx * 3.0 + offset, t * speed * 0.5]);
-                let n2 = self.noise.get([fx * 7.0 + offset + 10.0, t * speed * 0.3 + 5.0]);
+                let n2 = self
+                    .noise
+                    .get([fx * 7.0 + offset + 10.0, t * speed * 0.3 + 5.0]);
                 let curtain_y = h * (0.08 + n1 * 0.2 + n2 * 0.06);
-                let curtain_height = h * (0.35 + self.noise.get([fx * 2.0, t * 0.15 + offset]) * 0.2);
+                let curtain_height =
+                    h * (0.35 + self.noise.get([fx * 2.0, t * 0.15 + offset]) * 0.2);
 
                 for y in 0..canvas.height {
                     let fy = y as f64;
@@ -72,27 +77,35 @@ impl Animation for Aurora {
                     let (r, g, b) = match layer {
                         0 => {
                             // Bright green (classic aurora)
-                            ((20.0 * intensity) as u8,
-                             (180.0 + 75.0 * intensity) as u8,
-                             (40.0 + 60.0 * intensity) as u8)
+                            (
+                                (20.0 * intensity) as u8,
+                                (180.0 + 75.0 * intensity) as u8,
+                                (40.0 + 60.0 * intensity) as u8,
+                            )
                         }
                         1 => {
                             // Cyan-teal
-                            ((20.0 + 40.0 * intensity) as u8,
-                             (140.0 + 115.0 * intensity) as u8,
-                             (160.0 + 95.0 * intensity) as u8)
+                            (
+                                (20.0 + 40.0 * intensity) as u8,
+                                (140.0 + 115.0 * intensity) as u8,
+                                (160.0 + 95.0 * intensity) as u8,
+                            )
                         }
                         2 => {
                             // Purple-magenta
-                            ((120.0 + 100.0 * intensity) as u8,
-                             (20.0 + 50.0 * intensity) as u8,
-                             (160.0 + 95.0 * intensity) as u8)
+                            (
+                                (120.0 + 100.0 * intensity) as u8,
+                                (20.0 + 50.0 * intensity) as u8,
+                                (160.0 + 95.0 * intensity) as u8,
+                            )
                         }
                         _ => {
                             // Pink-red (rare top layer)
-                            ((180.0 + 75.0 * intensity) as u8,
-                             (30.0 + 40.0 * intensity) as u8,
-                             (80.0 + 80.0 * intensity) as u8)
+                            (
+                                (180.0 + 75.0 * intensity) as u8,
+                                (30.0 + 40.0 * intensity) as u8,
+                                (80.0 + 80.0 * intensity) as u8,
+                            )
                         }
                     };
 
@@ -105,9 +118,9 @@ impl Animation for Aurora {
                         // Blend colors when layers overlap
                         let (er, eg, eb) = canvas.colors[idx];
                         let t = intensity / existing_b.max(0.01);
-                        let mr = ((er as f64 * (1.0 - t) + r as f64 * t)) as u8;
-                        let mg = ((eg as f64 * (1.0 - t) + g as f64 * t)) as u8;
-                        let mb = ((eb as f64 * (1.0 - t) + b as f64 * t)) as u8;
+                        let mr = (er as f64 * (1.0 - t) + r as f64 * t) as u8;
+                        let mg = (eg as f64 * (1.0 - t) + g as f64 * t) as u8;
+                        let mb = (eb as f64 * (1.0 - t) + b as f64 * t) as u8;
                         canvas.set_colored(x, y, existing_b, mr, mg, mb);
                     }
                 }
